==> recsys/3_build_faiss_storage.ipynb <==
import faiss
import pickle
import torch
from tqdm import tqdm

Construct faiss index for authors

with open('embeddings/authors.pkl', 'rb') as f:
    author_embedding = pickle.load(f)

dim = author_embedding[0].shape[0]
k = 1000
quantiser = faiss.IndexFlatL2(dim) 
index = faiss.IndexIVFFlat(quantiser, dim, k)
vectors = torch.vstack(list(author_embedding.values())).numpy()

index = faiss.index_factory(dim, "IVF1000,Flat", faiss.METRIC_INNER_PRODUCT)
print(index.is_trained)   # False.
faiss.normalize_L2(vectors)
index.train(vectors)  # Train на нашем наборе векторов
 
# Обучение завершено, но векторов в индексе пока нет, так что добавляем их в индекс:
print(index.is_trained)  # True
print(index.ntotal)   # 0
index.add(vectors)
print(index.ntotal)   # 125931

topn = 20
index.nprobe = 10
D, I = index.search(vectors[:5], topn)  # Возвращает результат: Distances, Indices\
print(D)
print(I)

faiss.write_index(index, 'index/author.index')

Construct faiss index for papers

with open('embeddings/papers.pkl', 'rb') as f:
    paper_embeddings = pickle.load(f)

dim = 768
k = 1000
quantiser = faiss.IndexFlatL2(dim) 
index = faiss.IndexIVFFlat(quantiser, dim, k)
vectors = torch.vstack(list(paper_embeddings.values())).numpy()

index = faiss.index_factory(dim, "IVF1000,Flat", faiss.METRIC_INNER_PRODUCT)
print(index.is_trained)   # False.
faiss.normalize_L2(vectors)
index.train(vectors)  # Train на нашем наборе векторов
 
# Обучение завершено, но векторов в индексе пока нет, так что добавляем их в индекс:
print(index.is_trained)  # True
print(index.ntotal)   # 0
index.add(vectors)
print(index.ntotal)   # 51570

topn = 20
index.nprobe = 10
D, I = index.search(vectors[:5], topn)  # Возвращает результат: Distances, Indices\
print(D)
print(I)

faiss.write_index(index, 'index/paper.index')
