==> recsys/4_usage.ipynb <==
import faiss
import pickle

Usage: author recommendation

# load author mapping
with open('data/idx2author.pkl', 'rb') as f:
    idx2author = pickle.load(f)

with open('data/author2idx.pkl', 'rb') as f:
    author2idx = pickle.load(f)

# load embeddings
with open('embeddings/authors.pkl', 'rb') as f:
    author_embedding = pickle.load(f)

# load index
index = faiss.read_index('index/author.index')

author_name = 'Jan Aelterman'
topn = 5
index.nprobe = 10 # number of clusters to search
try: 
    idx = author2idx[author_name]
    embedding = author_embedding[idx].unsqueeze(0).numpy()
    faiss.normalize_L2(embedding)
    _, I = index.search(embedding, topn)
except KeyError:
    print('Author not in index')

for rec in I[0][1:]: # as first is always the same author
    print(idx2author[rec])

Usage: paper recommendation

# load embeddings
with open('embeddings/papers.pkl', 'rb') as f:
    paper_embeddings = pickle.load(f)

# load index
index = faiss.read_index('index/paper.index')

paper_id = '53e997ccb7602d9701fbe45d'
topn = 5
index.nprobe = 10 # number of clusters to search

embedding = paper_embeddings[paper_id].unsqueeze(0).numpy()
faiss.normalize_L2(embedding)
_, I = index.search(embedding, topn)


res_id = []
for rec in I[0]: # as first is always the same author
    id = list(paper_embeddings.keys())[rec]
    res_id.append(id)
    print(id)

import pandas as pd

data = pd.read_csv('../clustering/results/004_of_V13_topics.csv')
# first row - the paper for which we searched
print(data['abstract'][data['_id'].isin(res_id)])
